Update grid-world-generator.py
There are plenty of ways to generate mazes. Check this website https://en.wikipedia.org/wiki/Maze_generation_algorithm. I created a code "grid-world-generator.py" which implements two algorithms among them. To run the code
Python grid-world-generator.py <number of grids you specify>
One algorithm is called randGridMaze, where each cell is assigned to be 1 or 0 with certain probability.
The other algorithm is called backTrackerMaze, where the maze is generated with recursive backtracker. You can think of it as a squirrel carrying a shovel, digging in a fully-occupied maze.
The maze data are stored in the "arrs" folder and the pictures in "pics" folder.